re organize folder structure to group all navbar components
